Common module, kafka-properties, datetimeutils
Added a new common module with misc utility methods/classes that didn't seem right for the model (TBD whether it remains). Currently it contains some date time parsing utilities and several unit tests to exercise them.
Renamed the kafka-adapter's application.properties file to "kafka-application.properties" and added a "@PropertySource" on each of the modules that requires it, this will let us share common properties related to kafka (same idea will be employed elsewhere as needed)
Misc cleanup and clarification in the README
Tweaked the bootJar task to skip building a bootJar for library modules.